backend/requirements.txt
Django>=4.2
cairosvg
djangorestframework
djangorestframework-simplejwt
django-cors-headers
